// src/app/open-api/generated/model/file-transfer-update.ts
/**
 * GED API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface FileTransferUpdate { 
    /**
     * Unique ID of the file transfer
     */
    trackingId?: string;
    /**
     * Unique ID of the file transfer
     */
    trackingName?: string;
    /**
     * Name of the file being transferred
     */
    fileName?: string;
    /**
     * Total number of files in the transfer
     */
    totalFiles?: number;
    /**
     * Number of files transferred so far
     */
    transferredFiles?: number;
    /**
     * Total size of files being transferred in bytes
     */
    totalSize?: number;
    /**
     * Size of files transferred so far in bytes
     */
    transferredSize?: number;
    /**
     * Current status of the file transfer
     */
    status?: FileTransferUpdate.StatusEnum;
    /**
     * Type of flow for the file transfer
     */
    flowType?: FileTransferUpdate.FlowTypeEnum;
    /**
     * Optional message or status note
     */
    message?: string | null;
}
export namespace FileTransferUpdate {
    export const StatusEnum = {
        Pending: 'PENDING',
        InProgress: 'IN_PROGRESS',
        Completed: 'COMPLETED',
        Failed: 'FAILED',
        Cancelled: 'CANCELLED'
    } as const;
    export type StatusEnum = typeof StatusEnum[keyof typeof StatusEnum];
    export const FlowTypeEnum = {
        Import: 'import',
        Export: 'export'
    } as const;
    export type FlowTypeEnum = typeof FlowTypeEnum[keyof typeof FlowTypeEnum];
}